# api-gateway/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hire_Jobs is a platform to find your dream job.",
        "title": "Hireo Jobs API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/employer/company": {
            "get": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Retrieve details of the company deails with the authenticated employer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Fetch company details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company details retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid employer ID type",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch company details",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerTokenAuth": []
                    }
                ],
                "description": "Update details of the company associated with the authenticated employer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers"
                ],
                "summary": "Update company details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Employer details to update",
                        "name": "employerDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EmployerDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Company updated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid employer ID type",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Failed to update company",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/employer/login": {
            "post": {
                "description": "Process the employer login request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers Authentication"
                ],
                "summary": "Login for employer",
                "parameters": [
                    {
                        "description": "Employer Login details",
                        "name": "employer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EmployerLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employer authenticated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Details not in correct format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Cannot authenticate employer",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/employer/signup": {
            "post": {
                "description": "Process the employer signup request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employers Authentication"
                ],
                "summary": "Sign up a new employer",
                "parameters": [
                    {
                        "description": "Employer SignUp details",
                        "name": "employer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EmployerSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employer created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Details not in correct format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Cannot create employer",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/jobseeker/login": {
            "post": {
                "description": "Authenticate job seeker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Seekers Authentication"
                ],
                "summary": "Job seeker login",
                "parameters": [
                    {
                        "description": "Job seeker credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.JobSeekerLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job seeker authenticated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Details not in correct format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Cannot authenticate job seeker",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/jobseeker/signup": {
            "post": {
                "description": "Create a new job seeker account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Job Seekers Authentication"
                ],
                "summary": "Job seeker sign up",
                "parameters": [
                    {
                        "description": "Job seeker details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.JobSeekerSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job seeker created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Details not in correct format",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Cannot create job seeker",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.EmployerDetails": {
            "type": "object",
            "properties": {
                "about_company": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "company_size": {
                    "type": "integer"
                },
                "contact_email": {
                    "type": "string"
                },
                "contact_phone_number": {
                    "type": "integer"
                },
                "headquarters_address": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "models.EmployerLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
        "models.EmployerSignUp": {
            "type": "object",
            "properties": {
                "about_company": {
                    "type": "string"
                },
                "company_name": {
                    "type": "string"
                },
                "company_size": {
                    "type": "integer"
                },
                "contact_email": {
                    "type": "string"
                },
                "contact_phone_number": {
                    "type": "integer"
                },
                "headquarters_address": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "models.JobSeekerLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                }
            }
        },
        "models.JobSeekerSignUp": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 6
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerTokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}